Fixed Script, login and register swap working
