+++ src/app/login/login.component.ts
@@ -1,10 +1,30 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-login',
   templateUrl: './login.component.html',
-  styleUrl: './login.component.css'
+  styleUrls: ['./login.component.css']
 })
-export class LoginComponent {
+export class LoginComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+    const container = document.getElementById('container');
+    const registerBtn = document.getElementById('register');
+    const loginBtn = document.getElementById('login');
+
+    if (container && registerBtn && loginBtn) {
+      registerBtn.addEventListener('click', () => {
+        container.classList.add("active");
+      });
+
+      loginBtn.addEventListener('click', () => {
+        container.classList.remove("active");
+      });
+    } else {
+      console.error('Elements not found.');
+    }
+  }
 
 }
